tools/extract_cards/pyproject.toml
[project]
name = "extract-cards"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.1",
    "pillow>=11.2.1",
    "pypdfium2>=4.30.1",
]
